Refactor the template loader classes
All the template loader classes have been refactored to accept a
variable number of arguments. This shouldn't cause any backwards
compatibility problems.

The file system loader classes have been renamed. The old FastFileLoader
class is the new FileLoader class as it deserves to be the default file
system loader. The old FileLoader class is the new FileReloader class,
bumped from the default spot as automatic recompilation is likely to be
a fringe benefit in most use cases.

This represents a minor break in backwards compatibility, but the trade
off seems acceptable at this point.